% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae01.R
\docType{data}
\name{ae01}
\alias{ae01}
\title{Example AE data frame}
\format{
a data frame with N rows and M columns. columns include:
\describe{
\item{aebodsys}{body system organ class}
\item{trt}{treatment name}
\item{No_Event}{no event}
\item{Event}{Event }

}
}
\source{
its a sample data frame
}
\usage{
ae01
}
\description{
Example AE data frame
}
\examples{

head(ae01)
}
\keyword{datasets}
