% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_oddsrto.R
\name{compute_oddsrto}
\alias{compute_oddsrto}
\title{the compute_oddsrto is used to generate the data frame with odds ratio by adverse event}
\usage{
compute_oddsrto(.data, aeterm)
}
\arguments{
\item{.data}{pass any data frame}

\item{aeterm}{pass any ae column name}
}
\value{
a data frame
}
\description{
the compute_oddsrto is used to generate the data frame with odds ratio by adverse event
}
\examples{
compute_oddsrto(.data=ae01, aeterm='aebodsys')
}
